requirements.txt
beautifulsoup4>=4.10.0
edq-utils>=0.2.4
requests>=2.31.0
